Restructure settings folder
