+++ lib/screens/channel/chat/emote_menu/emote_menu_section.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:frosty/constants/constants.dart';
 import 'package:frosty/models/emotes.dart';
 import 'package:frosty/screens/channel/stores/chat_store.dart';
-import 'package:frosty/screens/settings/stores/settings_store.dart';
+import 'package:frosty/screens/settings/settings_store.dart';
 import 'package:provider/provider.dart';
 
 class EmoteMenuSection extends StatelessWidget {
