+++ lib/screens/settings/settings.dart
@@ -3,12 +3,12 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:frosty/core/auth/auth_store.dart';
-import 'package:frosty/screens/settings/sections/account_settings.dart';
-import 'package:frosty/screens/settings/sections/chat_settings.dart';
-import 'package:frosty/screens/settings/sections/general_settings.dart';
-import 'package:frosty/screens/settings/sections/other_settings.dart';
-import 'package:frosty/screens/settings/sections/video_settings.dart';
-import 'package:frosty/screens/settings/stores/settings_store.dart';
+import 'package:frosty/screens/settings/account_settings.dart';
+import 'package:frosty/screens/settings/chat_settings.dart';
+import 'package:frosty/screens/settings/general_settings.dart';
+import 'package:frosty/screens/settings/other_settings.dart';
+import 'package:frosty/screens/settings/settings_store.dart';
+import 'package:frosty/screens/settings/video_settings.dart';
 import 'package:provider/provider.dart';
 import 'package:url_launcher/url_launcher.dart';
